Prevent adding duplicate routes to sitemap
If a plugin makes a call to add a new route to the sitemap routes to be
included, it's possible that it will be adding a route which already
exists in the database and therefore will already have been included in
the sitemap content. The best way of preventing duplicates being created
in this manner is to just prevent the plugin api call from succeeding.